just cleaning up files, added connector class file and added to main.py and weather_vis file
